pyproject.toml
[project]
name = "webhook-inspector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Stanislas Plum", email = "57526019+quaxsze@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.1",
    "gunicorn>=26.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "opentelemetry-distro>=0.62b1",
    "opentelemetry-exporter-otlp>=1.41.1",
    "opentelemetry-instrumentation-fastapi>=0.62b1",
    "opentelemetry-instrumentation-psycopg>=0.62b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.62b1",
    "psycopg[binary]>=3.3.4",
    "pydantic-settings>=2.14.1",
    "sqlmodel>=0.0.38",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.46.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=2.1.0",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
    "testcontainers>=4.14.2",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "SIM"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
exclude = ["migrations/"]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
